+++ samples/Javascript/CocosPlayer/Classes/Java_org_cocos2dx_cocosplayer_CocosPlayerSocket.cpp
@@ -10,7 +10,6 @@
 #define  LOGD(...)  __android_log_print(ANDROID_LOG_DEBUG,LOG_TAG,__VA_ARGS__)
 
 #define  SOCKET_CLASS_NAME "org/cocos2dx/cocosplayer/CocosPlayerSocket"
-#define  HELPER_CLASS_NAME "org/cocos2dx/cocosplayer/CCBFileUtilsHelper"
 
 using namespace cocos2d;
 using namespace std;
@@ -26,7 +25,7 @@ extern "C" {
 
    void cleanCacheDirJNI() {
         JniMethodInfo t;
-        if (JniHelper::getStaticMethodInfo(t, HELPER_CLASS_NAME, "cleanCache", "()V")) {
+        if (JniHelper::getStaticMethodInfo(t, SOCKET_CLASS_NAME, "cleanCache", "()V")) {
 	  t.env->CallStaticVoidMethod(t.classID, t.methodID);
 	  t.env->DeleteLocalRef(t.classID);
         }
